Add SyntaxErrorInQueryAssemblerRule
This implements a new Rule for methods that build queries from simple
ingredients: a table name, and arrays of column name/value elements.
It checks that all table and column names are valid, and that the array
value types are (maybe) accepted by the column types.

For configuration, the class::method is specified as usual, with an
optional comma-delimited list of array agrument indices after the hash.
Specifying the indices is necessary because this rule covers 3 kinds of
simple builders:

1. No array arguments (for truncate-like queries)
2. One array argument (for insert/delete/replace-like queries)
3. Two array arguments (for update-like queries)

The most notable examples of this type of method are the Doctrine\DBAL
insert, delete, and update methods on their Connection object, which
have additional arguments that are not arrays of column names/values.